# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Skolara API
  version: 1.0.0
  description: |
    Skolara — The Intelligent Operating System for Schools.

    Conventions:
    - Versioned paths (`/api/v1/…`)
    - Consistent error envelope (`error.code` / `error.message` / `error.details`)
    - `X-Request-ID` propagated to logs and audit
    - `Idempotency-Key` supported on financial mutations (journal entries); payment
      webhooks are idempotent by provider event id
    - Authorization enforced server-side; tenant context derived from the
      verified session (X-School-ID of an ACTIVE membership), never from payloads
    - Amounts are INTEGER minor units (e.g. KES cents) — floats never
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Login with email+password (sets rotating refresh cookie)
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Access token issued (refresh token in HttpOnly cookie)
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
                  tokenType: { type: string, const: Bearer }
                  expiresIn: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "423": { $ref: "#/components/responses/Conflict" }
  /api/v1/auth/signup:
    post:
      tags: [auth]
      summary: Self-serve school onboarding — provisions admin user, education group, school, and membership in one transaction
      operationId: signup
      description: |
        Public (rate-limited) onboarding endpoint. Issues NO session — the
        client completes onboarding through the standard login flow, so
        cookie/session handling stays on one code path.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [schoolName, adminName, email, password]
              properties:
                schoolName: { type: string, minLength: 2, maxLength: 128 }
                adminName: { type: string, minLength: 1, maxLength: 120 }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        "201":
          description: Workspace provisioned (log in to obtain a session)
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id: { type: string }
                      email: { type: string }
                      name: { type: string }
                  school:
                    type: object
                    properties:
                      id: { type: string }
                      code: { type: string }
                      name: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Rotate refresh token (cookie or body) into a new access token
      operationId: refreshToken
      responses:
        "200": { description: New access + refresh tokens }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/me:
    get:
      tags: [auth]
      summary: Current user with effective permission set
      operationId: me
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserWithPermissions" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/users:
    get:
      tags: [users]
      summary: List users (paginated)
      operationId: listUsers
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: User page
          headers: { X-Total-Count: { $ref: "#/components/headers/TotalCount" } }
          content:
            application/json:
              schema:
                type: object
                properties:
                  users: { type: array, items: { $ref: "#/components/schemas/User" } }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [users]
      summary: Provision a user (admin flow)
      operationId: createUser
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name, password]
              properties:
                email: { type: string }
                name: { type: string }
                password: { type: string, minLength: 8 }
                role: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/users/{id}:
    get:
      tags: [users]
      summary: Fetch a user
      operationId: getUser
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      responses:
        "200":
          description: User
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/users/{id}/roles:
    post:
      tags: [users]
      summary: Grant a platform role to a user
      operationId: assignRole
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string }
      responses:
        "204": { description: Assigned }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/users/{id}/status:
    patch:
      tags: [users]
      summary: Enable/disable an account (disabling revokes all sessions)
      operationId: setUserStatus
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [active, disabled] }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  headers:
    TotalCount:
      description: Total number of records matching the query
      schema: { type: integer }
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: Missing permission or school context
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: Resource not found (also used cross-tenant to avoid enumeration)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: Domain conflict (state machine, uniqueness, overlap)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  issue: { type: string }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        name: { type: string }
        status: { type: string, enum: [active, disabled, locked] }
    UserWithPermissions:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            roles: { type: array, items: { type: string } }
            permissions:
              type: object
              additionalProperties: { type: boolean }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT